LayoutTests/platform/chromium/media/controls-audio-sizes.html

<!DOCTYPE html>
<html>
<head>
    <title>Audio Controls - selective control element display</title>
    <script src="../../../media/video-test.js"></script>
    <script src="../../../media/media-file.js"></script>
    <script src="../../../media/media-controls.js"></script>
    <script>
    var audios = document.getElementsByTagName("audio");
    var ready = 0;

    function test(audioId)
    {
        var audio = audios[audioId];
        var ready = 0;

        if (window.internals) {
            var shadow = internals.shadowRoot(audio).firstChild;
            var width = audio.clientWidth;

            enclosure     = mediaControlsElement(shadow, "-webkit-media-controls-enclosure");
            remainingtime = mediaControlsElement(shadow, "-webkit-media-controls-time-remaining-display");
            volumeslider  = mediaControlsElement(shadow, "-webkit-media-controls-volume-slider");
            mutebutton    = mediaControlsElement(shadow, "-webkit-media-controls-mute-button");
            timeline      = mediaControlsElement(shadow, "-webkit-media-controls-timeline");

            consoleWrite("");
            consoleWrite("** The audio width is " + width + " **");
            testExpected("getComputedStyle(enclosure)['display']", 'block', "==");

            if (width >= 350)
                testExpected("getComputedStyle(remainingtime)['display']", 'block', "==");
            else
                testExpected("getComputedStyle(remainingtime)['display']", 'none', "==");

            if (width >= 275)
                testExpected("getComputedStyle(volumeslider)['display']", '-webkit-box', "==");
            else
                testExpected("getComputedStyle(volumeslider)['display']", 'none', "==");

            if (width >= 210)
                testExpected("getComputedStyle(mutebutton)['display']", '-webkit-box', "==");
            else
                testExpected("getComputedStyle(mutebutton)['display']", 'none', "==");

            if (width >= 100)
                testExpected("getComputedStyle(timeline)['display']", '-webkit-box', "==");
            else
                testExpected("getComputedStyle(timeline)['display']", 'none', "==");
        }
        var next = audioId + 1;
        if (next == audios.length)
            endTest();
        else
            test(next);
    }

    function init()
    {
        var file = findMediaFile('audio', '../../../media/content/short');
        for (var i = 0; i < audios.length; i++) {
            audios[i].src = file;
            audios[i].addEventListener('canplaythrough', function() {
                ready += 1;
                if (ready == audios.length) {
                    test(0);
                }
            }, false);
        }
    }
    </script>
</head>
<body onload="init()">
    <p>Lose time display at 349px in Chromium</p>
    <audio controls style="width: 350px;"></audio>
    <audio controls style="width: 349px;"></audio>
    <p>Lose volume slider at 274px in Chromium</p>
    <audio controls style="width: 275px;"></audio>
    <audio controls style="width: 274px;"></audio>
    <p>Lose mute button at 209px in Chromium</p>
    <audio controls style="width: 210px;"></audio>
    <audio controls style="width: 209px;"></audio>
    <p>Lose transport bar at 99px in Chromium</p>
    <audio controls style="width: 100px;"></audio>
    <audio controls style="width: 99px;"></audio>
</body>
</html>
